refactor: extract shared client logic into _RegistryClientBase
Move CachedSchema, closed-client guard, int64 validation, endpoint
building, and response processing into a shared base class. Both
sync and async clients now inherit from _RegistryClientBase,
eliminating code duplication that previously caused divergence bugs.
